backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
sqlalchemy
asyncpg
greenlet